requirements.txt
# Flask Core
Flask==3.0.3
Werkzeug==3.0.3

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Flask-Mail==0.10.0

# Database
alembic==1.13.2

# Environment & Configuration
python-dotenv==1.0.1

# HTTP & API Clients
requests==2.32.3

# IP & AS Lookup
pyasn==1.6.2
ipwhois==1.3.0
maxminddb==2.6.2

# Validation
email-validator==2.2.0
validators==0.33.0
dnspython==2.6.1

# WSGI Server (Production)
gunicorn==22.0.0

# Security
cryptography==42.0.8

# Utilities
python-dateutil==2.9.0

# Markdown & Content
markdown==3.6
bleach==6.1.0
pymdown-extensions==10.8.1
